// src/app/modules/user/user.interface.ts
// step1: creating an interface
export interface IUser{
    id: string
    role: 'student'
    password: string
    name: {
        firstName: string
        middleName?: string
        lastname: string
    }
    dateOfBirth?: string
    gender: "male" | 'female'
    email?: string
    contactNo: string
    emergencyContactNo: string
    presentAddress: string
    permanentAddress: string
}